Create windows on demand, not at startup
Previously, a window was created at startup for each display, even if
the display was not enabled. This resulted in a fixed 1:1 association
between windows and remote displays. Since there was always one window
created at startup to display status messages (the "main window"), this
was always associated with remote display #1. But if the first remote
display was not enabled, we ended up with a extra black window with a
message saying ("Waiting for display 1...").

By creating windows on demand, we can re-use the "main window" for any
arbitrary display, even if it's not display #1.

Resolves: rhbz#1032939